Enum fix

+++ DeliveryService/Controllers/ProductController.cs
@@ -24,7 +24,7 @@ namespace DeliveryService.Controllers
 
         public IList<Product> GetProducts()
         {
-            return (List<Product>)_cache.GetFromCache((int)Cache.CollectionType.Product, () => _storeContext.Products);
+            return (List<Product>)_cache.GetFromCache(Cache.CollectionType.Product, () => _storeContext.Products);
         }
 
         public void AddProduct(Product product)
